Calculate a suite of long-term and seasonal flow statistics for one or more sites.

calc_flowstats uses a time series of measured or modeled flows to calculate a suite of long-term and seasonal (winter and summer) flow statistics for one or more sites (stations). It is primarily designed to work with mean daily flows (e.g. as produced by import_flow), but can also be applied to time series data on a longer (e.g. monthly) time step. Regardless, the data should be regularly spaced and the same time step should be used for all sites.

Usage

calc_flowstats_old(data = data, site_col = "flow_site_id", date_col = "date", flow_col = "flow", ref_col = NULL)

Arguments

data

Name of data frame or tibble containing the flow data to be processed. Must be in long format and contain columns for flow site id, date and flow (see output from import_flows).

date_col

Name of column in 'data' containing date of flow record. Default = "date".

flow_col

Name of column in 'data' containing flow data for processing. Default = "flow".

ref_col

Name of column in 'data' containing flow data for reference scenario against which flow statistics are standardised. Default: NULL.

sites_col

Name of column in 'data' containing unique flow site id. Default = "flow_site_id".

Value

calc_flowstats returns a list of two data frames. The first data frame contains a suite of time-varying flow statistics for every 6 month winter/summer period at every site. The columns are as follows:

flow_site_id (a unique site id; this is the standardised column header for flow sites)

season (S = summer, W = winter)

water_year (year starting 1st October; e.g. 2018 = 01/10/2018 to 30/09/2019)

Q10: the unstandardised Q10 flow in a 6-month winter or summer period.

Q30: the unstandardised Q30 flow in a 6-month winter or summer period.

Q50: the unstandardised Q50 flow in a 6-month winter or summer period.

Q70: the unstandardised Q70 flow in a 6-month winter or summer period.

Q90: the unstandardised Q70 flow in a 6-month winter or summer period.

Q95: the unstandardised Q95 flow in a 6-month winter or summer period.

Q99: the unstandardised Q99 flow in a 6-month winter or summer period.

Q10mean: the seasonal (winter or summer) Q10 flow, averaged across years.

Q10sd: the standard deviation of the seasonal (winter or summer) Q10 flows.

Q10z: the Q10 flow in a six month period, standardised using the above mean and sd. If ref_col = NULL, then the mean and sd parameters are calculated using the same flow time series (i.e. Q10z = (Q10 - Q10mean) / Q10sd); if not, the mean and sd parameters are calculated from the ref_col flow time series (i.e. Q10z = (Q10 - Q10mean_ref) / Q10sd_ref).

Q30mean: (as for Q10)

Q30sd: (as for Q10)

Q30z: (as for Q10)

Q50mean: (as for Q10)

Q50sd: (as for Q10)

Q50z: (as for Q10)

Q70mean: (as for Q10)

Q70sd: (as for Q10)

Q70z: (as for Q10)

Q95: (as for Q10)

Q95mean: (as for Q10)

Q95sd: (as for Q10)

Q95z: (as for Q10)

zero: Number of zero flow days

mean: mean flow

sd: standard deviation of flows

lsd: standard deviation of the log flow series

n: number of days (182 or 183)

e3: number of events when flows exceed 3 x the long-term median (Q50) flow

e5: number of events when flows exceed 5 x the long-term median (Q50) flow

e7: number of events when flows exceed 7 x the long-term median (Q50) flow

missing: number of missing records (NAs)

durationAbove: number of records above the long-term Q70

nEventsAbove: number of events when flows exceed the long-term Q70

durationBelow: number of records below the long-term Q95

nEventsBelow: number of events when flows exceed the long-term Q95

e3mean: e3, averaged across years

e5mean: e5, averaged across years

e7mean: e7, averaged across years

durationAboveMean: durationAbove, averaged across years

durationBelowMean: durationBelow, averaged across years

nEventsAboveMean: nEventsAbove, averaged across years

nEventsBelowMean: nEventsBelow, averaged across years

The second data table contains long-term flow statistics. The data are arranged in long format, with the following columns:

flow_site_id (a unique site id)

season (S = summer, W = winter, A = all)

parameter (base flow index (bfi), flow duration curve percentiles (p1 to p99), and seasonal means and standard deviations for Q10, Q30, Q50, Q75 and Q95).

value (calculated statistic)

Details

The function is hard-wired to exclude any flow data before 1st April 1985. Must be in long format and contain columns named "flow_site_id", "Date" and "Flow" (see output from import_flows).

Winter is defined as 1st October to 31st March, and Summer as 1st April to 30th September. There is no minimum number of records required in each 6 month period; if desired, the outputs can be filtered afterwards to eliminate potentially unreliable results based on sparse data.

Selected flow percentiles (Q10, Q30, Q50, Q70 and Q95) are calculated for each 6 month winter/summer, and then standardized using the long-term mean and standard deviation of these statistics. If calculating flow statistics for two different flow scenarios (e.g. historical and naturalised), then the function includes the facility to standardise the flow percentile statistics for one scenario (specified via flow_col) using mean and standard deviation of flow statistics from the other scenario (specified via ref_col). For example, if flow_col = naturalised flows and ref_col = historical flows, then the resulting statistics can be input into a hydro-ecological model calibrated using historical flow data and used to make predictions of ecological status under naturalised flows.

Examples

## Calculate a suite of long-term and seasonal flow statistics for all sites listed in site_col:
calc_flowstats(data = flow_data,
                site_col = "flow_site_id",
                date_col = "date",
                flow_col = "flow")
#> |==============================================        | 86% ~0 s remaining     |===============================================       | 88% ~0 s remaining     |================================================      | 90% ~0 s remaining     |==================================================    | 93% ~0 s remaining     |===================================================   | 95% ~0 s remaining     |====================================================  | 97% ~0 s remaining     |===================================================== | 99% ~0 s remaining     
#> Joining with `by = join_by(site, win_no, n_data)`
#> |============                                          | 23% ~7 s remaining     |============                                          | 24% ~7 s remaining     |=============                                         | 25% ~7 s remaining     |=============                                         | 25% ~7 s remaining     |=============                                         | 26% ~6 s remaining     |==============                                        | 26% ~6 s remaining     |==============                                        | 27% ~6 s remaining     |==============                                        | 28% ~6 s remaining     |===============                                       | 28% ~6 s remaining     |===============                                       | 29% ~6 s remaining     |================                                      | 30% ~6 s remaining     |================                                      | 30% ~6 s remaining     |================                                      | 31% ~6 s remaining     |=================                                     | 32% ~6 s remaining     |=================                                     | 32% ~6 s remaining     |=================                                     | 33% ~6 s remaining     |==================                                    | 34% ~6 s remaining     |==================                                    | 34% ~6 s remaining     |==================                                    | 35% ~6 s remaining     |===================                                   | 36% ~6 s remaining     |===================                                   | 36% ~6 s remaining     |===================                                   | 37% ~5 s remaining     |====================                                  | 38% ~5 s remaining     |====================                                  | 38% ~5 s remaining     |====================                                  | 39% ~5 s remaining     |=====================                                 | 39% ~5 s remaining     |=====================                                 | 40% ~5 s remaining     |=====================                                 | 41% ~5 s remaining     |======================                                | 41% ~5 s remaining     |======================                                | 42% ~5 s remaining     |=======================                               | 43% ~5 s remaining     |=======================                               | 44% ~5 s remaining     |========================                              | 45% ~5 s remaining     |========================                              | 45% ~5 s remaining     |========================                              | 46% ~5 s remaining     |=========================                             | 47% ~5 s remaining     |=========================                             | 48% ~5 s remaining     |==========================                            | 48% ~4 s remaining     |==========================                            | 49% ~4 s remaining     |==========================                            | 50% ~4 s remaining     |===========================                           | 50% ~4 s remaining     |===========================                           | 51% ~4 s remaining     |===========================                           | 52% ~4 s remaining     |============================                          | 52% ~4 s remaining     |============================                          | 53% ~4 s remaining     |============================                          | 53% ~4 s remaining     |=============================                         | 54% ~4 s remaining     |=============================                         | 55% ~4 s remaining     |==============================                        | 56% ~4 s remaining     |==============================                        | 56% ~4 s remaining     |==============================                        | 57% ~4 s remaining     |===============================                       | 58% ~4 s remaining     |===============================                       | 59% ~4 s remaining     |================================                      | 59% ~4 s remaining     |================================                      | 60% ~3 s remaining     |================================                      | 60% ~3 s remaining     |=================================                     | 61% ~3 s remaining     |=================================                     | 62% ~3 s remaining     |=================================                     | 63% ~3 s remaining     |==================================                    | 63% ~3 s remaining     |==================================                    | 64% ~3 s remaining     |==================================                    | 65% ~3 s remaining     |===================================                   | 65% ~3 s remaining     |===================================                   | 66% ~3 s remaining     |====================================                  | 67% ~3 s remaining     |====================================                  | 67% ~3 s remaining     |====================================                  | 68% ~3 s remaining     |=====================================                 | 69% ~3 s remaining     |=====================================                 | 69% ~3 s remaining     |=====================================                 | 70% ~3 s remaining     |======================================                | 71% ~3 s remaining     |======================================                | 72% ~2 s remaining     |=======================================               | 73% ~2 s remaining     |=======================================               | 74% ~2 s remaining     |========================================              | 75% ~2 s remaining     |========================================              | 75% ~2 s remaining     |=========================================             | 76% ~2 s remaining     |=========================================             | 77% ~2 s remaining     |==========================================            | 78% ~2 s remaining     |==========================================            | 79% ~2 s remaining     |==========================================            | 80% ~2 s remaining     |===========================================           | 80% ~2 s remaining     |===========================================           | 81% ~2 s remaining     |============================================          | 82% ~2 s remaining     |============================================          | 83% ~1 s remaining     |=============================================         | 84% ~1 s remaining     |=============================================         | 84% ~1 s remaining     |=============================================         | 85% ~1 s remaining     |==============================================        | 86% ~1 s remaining     |==============================================        | 87% ~1 s remaining     |===============================================       | 88% ~1 s remaining     |===============================================       | 88% ~1 s remaining     |================================================      | 89% ~1 s remaining     |================================================      | 90% ~1 s remaining     |=================================================     | 91% ~1 s remaining     |=================================================     | 92% ~1 s remaining     |==================================================    | 93% ~1 s remaining     |==================================================    | 93% ~1 s remaining     |==================================================    | 94% ~0 s remaining     |===================================================   | 95% ~0 s remaining     |===================================================   | 95% ~0 s remaining     |====================================================  | 96% ~0 s remaining     |====================================================  | 97% ~0 s remaining     |====================================================  | 98% ~0 s remaining     |===================================================== | 99% ~0 s remaining     |===================================================== |100% ~0 s remaining     |==================                                    | 34% ~4 s remaining     |==================                                    | 35% ~4 s remaining     |===================                                   | 36% ~4 s remaining     |===================                                   | 37% ~4 s remaining     |====================                                  | 38% ~4 s remaining     |=====================                                 | 39% ~4 s remaining     |=====================                                 | 40% ~3 s remaining     |======================                                | 41% ~3 s remaining     |======================                                | 42% ~3 s remaining     |=======================                               | 43% ~3 s remaining     |=======================                               | 44% ~3 s remaining     |========================                              | 45% ~3 s remaining     |========================                              | 46% ~3 s remaining     |=========================                             | 47% ~3 s remaining     |=========================                             | 48% ~3 s remaining     |==========================                            | 49% ~3 s remaining     |===========================                           | 50% ~3 s remaining     |===========================                           | 51% ~3 s remaining     |============================                          | 52% ~3 s remaining     |============================                          | 53% ~3 s remaining     |=============================                         | 54% ~3 s remaining     |=============================                         | 55% ~3 s remaining     |==============================                        | 56% ~3 s remaining     |==============================                        | 57% ~2 s remaining     |===============================                       | 58% ~2 s remaining     |===============================                       | 59% ~2 s remaining     |================================                      | 60% ~2 s remaining     |================================                      | 61% ~2 s remaining     |=================================                     | 62% ~2 s remaining     |=================================                     | 63% ~2 s remaining     |==================================                    | 64% ~2 s remaining     |==================================                    | 65% ~2 s remaining     |===================================                   | 66% ~2 s remaining     |====================================                  | 67% ~2 s remaining     |====================================                  | 68% ~2 s remaining     |=====================================                 | 69% ~2 s remaining     |=====================================                 | 70% ~2 s remaining     |======================================                | 71% ~2 s remaining     |======================================                | 72% ~2 s remaining     |=======================================               | 73% ~2 s remaining     |=======================================               | 74% ~1 s remaining     |========================================              | 75% ~1 s remaining     |=========================================             | 76% ~1 s remaining     |=========================================             | 77% ~1 s remaining     |==========================================            | 78% ~1 s remaining     |==========================================            | 79% ~1 s remaining     |===========================================           | 80% ~1 s remaining     |===========================================           | 81% ~1 s remaining     |============================================          | 82% ~1 s remaining     |============================================          | 83% ~1 s remaining     |=============================================         | 84% ~1 s remaining     |=============================================         | 85% ~1 s remaining     |==============================================        | 86% ~1 s remaining     |===============================================       | 87% ~1 s remaining     |===============================================       | 88% ~1 s remaining     |===============================================       | 89% ~1 s remaining     |================================================      | 90% ~1 s remaining     |=================================================     | 91% ~0 s remaining     |=================================================     | 92% ~0 s remaining     |==================================================    | 93% ~0 s remaining     |==================================================    | 94% ~0 s remaining     |===================================================   | 95% ~0 s remaining     |===================================================   | 96% ~0 s remaining     |====================================================  | 97% ~0 s remaining     |====================================================  | 98% ~0 s remaining     |===================================================== | 99% ~0 s remaining     |===========================                           | 50% ~2 s remaining     |===========================                           | 51% ~2 s remaining     |============================                          | 52% ~2 s remaining     |============================                          | 53% ~2 s remaining     |=============================                         | 54% ~2 s remaining     |=============================                         | 55% ~2 s remaining     |==============================                        | 56% ~2 s remaining     |==============================                        | 57% ~2 s remaining     |===============================                       | 58% ~2 s remaining     |================================                      | 59% ~2 s remaining     |================================                      | 60% ~2 s remaining     |=================================                     | 61% ~2 s remaining     |=================================                     | 62% ~2 s remaining     |==================================                    | 64% ~2 s remaining     |==================================                    | 64% ~2 s remaining     |===================================                   | 65% ~1 s remaining     |===================================                   | 66% ~1 s remaining     |====================================                  | 68% ~1 s remaining     |=====================================                 | 69% ~1 s remaining     |=====================================                 | 70% ~1 s remaining     |======================================                | 71% ~1 s remaining     |======================================                | 72% ~1 s remaining     |=======================================               | 73% ~1 s remaining     |=======================================               | 73% ~1 s remaining     |========================================              | 75% ~1 s remaining     |========================================              | 76% ~1 s remaining     |==========================================            | 79% ~1 s remaining     |===========================================           | 81% ~1 s remaining     |============================================          | 82% ~1 s remaining     |=============================================         | 83% ~1 s remaining     |=============================================         | 84% ~1 s remaining     |==============================================        | 86% ~1 s remaining     |===============================================       | 87% ~1 s remaining     |================================================      | 90% ~0 s remaining     |================================================      | 91% ~0 s remaining     |=================================================     | 91% ~0 s remaining     |==================================================    | 93% ~0 s remaining     |==================================================    | 94% ~0 s remaining     |===================================================   | 95% ~0 s remaining     |===================================================   | 96% ~0 s remaining     |===================================================== | 99% ~0 s remaining     |=============                                         | 25% ~6 s remaining     |==============                                        | 26% ~6 s remaining     |==============                                        | 26% ~6 s remaining     |==============                                        | 27% ~6 s remaining     |===============                                       | 28% ~6 s remaining     |===============                                       | 29% ~6 s remaining     |===============                                       | 30% ~6 s remaining     |================                                      | 30% ~6 s remaining     |================                                      | 31% ~6 s remaining     |================                                      | 31% ~5 s remaining     |=================                                     | 32% ~5 s remaining     |=================                                     | 33% ~5 s remaining     |==================                                    | 34% ~5 s remaining     |==================                                    | 35% ~5 s remaining     |===================                                   | 35% ~5 s remaining     |===================                                   | 36% ~5 s remaining     |===================                                   | 37% ~5 s remaining     |====================                                  | 37% ~5 s remaining     |====================                                  | 38% ~5 s remaining     |====================                                  | 39% ~5 s remaining     |=====================                                 | 40% ~5 s remaining     |=====================                                 | 40% ~5 s remaining     |=====================                                 | 41% ~5 s remaining     |======================                                | 42% ~5 s remaining     |======================                                | 42% ~5 s remaining     |=======================                               | 43% ~5 s remaining     |=======================                               | 44% ~5 s remaining     |========================                              | 45% ~4 s remaining     |========================                              | 45% ~4 s remaining     |========================                              | 46% ~4 s remaining     |=========================                             | 46% ~4 s remaining     |=========================                             | 47% ~4 s remaining     |=========================                             | 48% ~4 s remaining     |==========================                            | 49% ~4 s remaining     |==========================                            | 49% ~4 s remaining     |==========================                            | 50% ~4 s remaining     |===========================                           | 50% ~4 s remaining     |===========================                           | 51% ~4 s remaining     |============================                          | 52% ~4 s remaining     |============================                          | 53% ~4 s remaining     |=============================                         | 54% ~4 s remaining     |=============================                         | 55% ~4 s remaining     |=============================                         | 55% ~4 s remaining     |==============================                        | 56% ~4 s remaining     |==============================                        | 56% ~4 s remaining     |==============================                        | 57% ~3 s remaining     |===============================                       | 58% ~3 s remaining     |===============================                       | 59% ~3 s remaining     |================================                      | 59% ~3 s remaining     |================================                      | 60% ~3 s remaining     |================================                      | 61% ~3 s remaining     |=================================                     | 61% ~3 s remaining     |=================================                     | 62% ~3 s remaining     |=================================                     | 63% ~3 s remaining     |==================================                    | 64% ~3 s remaining     |==================================                    | 64% ~3 s remaining     |===================================                   | 65% ~3 s remaining     |===================================                   | 66% ~3 s remaining     |====================================                  | 67% ~3 s remaining     |====================================                  | 68% ~3 s remaining     |====================================                  | 68% ~3 s remaining     |=====================================                 | 69% ~2 s remaining     |=====================================                 | 70% ~2 s remaining     |=====================================                 | 70% ~2 s remaining     |======================================                | 71% ~2 s remaining     |======================================                | 71% ~2 s remaining     |======================================                | 72% ~2 s remaining     |=======================================               | 73% ~2 s remaining     |=======================================               | 73% ~2 s remaining     |=======================================               | 74% ~2 s remaining     |========================================              | 74% ~2 s remaining     |========================================              | 75% ~2 s remaining     |========================================              | 75% ~2 s remaining     |=========================================             | 76% ~2 s remaining     |=========================================             | 77% ~2 s remaining     |=========================================             | 77% ~2 s remaining     |==========================================            | 78% ~2 s remaining     |==========================================            | 79% ~2 s remaining     |==========================================            | 79% ~2 s remaining     |===========================================           | 80% ~2 s remaining     |===========================================           | 81% ~2 s remaining     |===========================================           | 81% ~2 s remaining     |============================================          | 82% ~1 s remaining     |============================================          | 83% ~1 s remaining     |=============================================         | 83% ~1 s remaining     |=============================================         | 84% ~1 s remaining     |=============================================         | 85% ~1 s remaining     |==============================================        | 85% ~1 s remaining     |==============================================        | 86% ~1 s remaining     |==============================================        | 87% ~1 s remaining     |===============================================       | 87% ~1 s remaining     |===============================================       | 88% ~1 s remaining     |===============================================       | 89% ~1 s remaining     |================================================      | 90% ~1 s remaining     |================================================      | 90% ~1 s remaining     |================================================      | 91% ~1 s remaining     |=================================================     | 91% ~1 s remaining     |=================================================     | 92% ~1 s remaining     |==================================================    | 93% ~1 s remaining     |==================================================    | 93% ~1 s remaining     |==================================================    | 94% ~0 s remaining     |===================================================   | 95% ~0 s remaining     |===================================================   | 95% ~0 s remaining     |===================================================   | 96% ~0 s remaining     |====================================================  | 97% ~0 s remaining     |====================================================  | 98% ~0 s remaining     |===================================================== | 98% ~0 s remaining     |===================================================== | 99% ~0 s remaining     |===================================================== |100% ~0 s remaining     |=============                                         | 25% ~6 s remaining     |=============                                         | 26% ~6 s remaining     |==============                                        | 26% ~6 s remaining     |==============                                        | 27% ~6 s remaining     |===============                                       | 28% ~6 s remaining     |===============                                       | 29% ~6 s remaining     |================                                      | 30% ~6 s remaining     |================                                      | 30% ~6 s remaining     |================                                      | 31% ~5 s remaining     |=================                                     | 32% ~5 s remaining     |=================                                     | 33% ~5 s remaining     |==================                                    | 33% ~5 s remaining     |==================                                    | 34% ~5 s remaining     |==================                                    | 35% ~5 s remaining     |===================                                   | 36% ~5 s remaining     |===================                                   | 37% ~5 s remaining     |====================                                  | 37% ~5 s remaining     |====================                                  | 38% ~5 s remaining     |====================                                  | 39% ~5 s remaining     |=====================                                 | 40% ~5 s remaining     |=====================                                 | 40% ~5 s remaining     |======================                                | 41% ~5 s remaining     |======================                                | 42% ~5 s remaining     |=======================                               | 43% ~5 s remaining     |=======================                               | 44% ~4 s remaining     |=======================                               | 44% ~4 s remaining     |========================                              | 45% ~4 s remaining     |========================                              | 45% ~4 s remaining     |========================                              | 46% ~4 s remaining     |=========================                             | 47% ~4 s remaining     |=========================                             | 47% ~4 s remaining     |==========================                            | 48% ~4 s remaining     |==========================                            | 49% ~4 s remaining     |==========================                            | 50% ~4 s remaining     |===========================                           | 50% ~4 s remaining     |===========================                           | 51% ~4 s remaining     |============================                          | 52% ~4 s remaining     |============================                          | 53% ~4 s remaining     |============================                          | 53% ~4 s remaining     |=============================                         | 54% ~4 s remaining     |=============================                         | 55% ~4 s remaining     |=============================                         | 55% ~4 s remaining     |==============================                        | 56% ~3 s remaining     |==============================                        | 57% ~3 s remaining     |===============================                       | 58% ~3 s remaining     |===============================                       | 59% ~3 s remaining     |================================                      | 60% ~3 s remaining     |================================                      | 60% ~3 s remaining     |================================                      | 61% ~3 s remaining     |=================================                     | 61% ~3 s remaining     |=================================                     | 62% ~3 s remaining     |==================================                    | 63% ~3 s remaining     |==================================                    | 64% ~3 s remaining     |==================================                    | 64% ~3 s remaining     |==================================                    | 65% ~3 s remaining     |===================================                   | 66% ~3 s remaining     |===================================                   | 66% ~3 s remaining     |====================================                  | 67% ~3 s remaining     |====================================                  | 68% ~3 s remaining     |=====================================                 | 69% ~3 s remaining     |=====================================                 | 69% ~2 s remaining     |=====================================                 | 70% ~2 s remaining     |======================================                | 71% ~2 s remaining     |======================================                | 71% ~2 s remaining     |======================================                | 72% ~2 s remaining     |=======================================               | 73% ~2 s remaining     |=======================================               | 74% ~2 s remaining     |========================================              | 74% ~2 s remaining     |========================================              | 75% ~2 s remaining     |========================================              | 76% ~2 s remaining     |=========================================             | 76% ~2 s remaining     |=========================================             | 77% ~2 s remaining     |==========================================            | 78% ~2 s remaining     |==========================================            | 79% ~2 s remaining     |==========================================            | 80% ~2 s remaining     |===========================================           | 80% ~2 s remaining     |===========================================           | 81% ~2 s remaining     |===========================================           | 81% ~1 s remaining     |============================================          | 82% ~1 s remaining     |============================================          | 83% ~1 s remaining     |=============================================         | 83% ~1 s remaining     |=============================================         | 84% ~1 s remaining     |=============================================         | 85% ~1 s remaining     |==============================================        | 85% ~1 s remaining     |==============================================        | 86% ~1 s remaining     |==============================================        | 87% ~1 s remaining     |===============================================       | 88% ~1 s remaining     |===============================================       | 89% ~1 s remaining     |================================================      | 89% ~1 s remaining     |================================================      | 90% ~1 s remaining     |================================================      | 91% ~1 s remaining     |=================================================     | 91% ~1 s remaining     |=================================================     | 92% ~1 s remaining     |==================================================    | 93% ~1 s remaining     |==================================================    | 94% ~1 s remaining     |==================================================    | 94% ~0 s remaining     |===================================================   | 95% ~0 s remaining     |===================================================   | 95% ~0 s remaining     |===================================================   | 96% ~0 s remaining     |====================================================  | 97% ~0 s remaining     |====================================================  | 98% ~0 s remaining     |===================================================== | 98% ~0 s remaining     |===================================================== | 99% ~0 s remaining     |===================================================== |100% ~0 s remaining     
#> [[1]]
#>    flow_site_id win_no start_date   end_date n_data n_missing n_total
#> 1        031031      1 1995-04-01 1995-09-30      0       183     183
#> 2        031031      2 1995-10-01 1996-03-31      0       183     183
#> 3        031031      3 1996-04-01 1996-09-30      0       183     183
#> 4        031031      4 1996-10-01 1997-03-31      0       182     182
#> 5        031031      5 1997-04-01 1997-09-30      0       183     183
#> 6        031031      6 1997-10-01 1998-03-31      0       182     182
#> 7        031031      7 1998-04-01 1998-09-30      0       183     183
#> 8        031031      8 1998-10-01 1999-03-31      0       182     182
#> 9        031031      9 1999-04-01 1999-09-30      0       183     183
#> 10       031031     10 1999-10-01 2000-03-31      0       183     183
#> 11       031031     11 2000-04-01 2000-09-30      0       183     183
#> 12       031031     12 2000-10-01 2001-03-31      0       182     182
#> 13       031031     13 2001-04-01 2001-09-30      0       183     183
#> 14       031031     14 2001-10-01 2002-03-31      0       182     182
#>    prop_missing mean sd Q5 Q10 Q20 Q25 Q30 Q50 Q70 Q75 Q80 Q90 Q95 Q99 Q5z Q10z
#> 1             1   NA NA NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA   NA
#> 2             1   NA NA NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA   NA
#> 3             1   NA NA NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA   NA
#> 4             1   NA NA NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA   NA
#> 5             1   NA NA NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA   NA
#> 6             1   NA NA NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA   NA
#> 7             1   NA NA NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA   NA
#> 8             1   NA NA NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA   NA
#> 9             1   NA NA NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA   NA
#> 10            1   NA NA NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA   NA
#> 11            1   NA NA NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA   NA
#> 12            1   NA NA NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA   NA
#> 13            1   NA NA NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA   NA
#> 14            1   NA NA NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  NA   NA
#>    Q20z Q25z Q30z Q50z Q70z Q75z Q80z Q90z Q95z Q99z dry_n dry_e dry_start
#> 1    NA   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA        NA
#> 2    NA   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA        NA
#> 3    NA   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA        NA
#> 4    NA   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA        NA
#> 5    NA   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA        NA
#> 6    NA   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA        NA
#> 7    NA   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA        NA
#> 8    NA   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA        NA
#> 9    NA   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA        NA
#> 10   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA        NA
#> 11   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA        NA
#> 12   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA        NA
#> 13   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA        NA
#> 14   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA        NA
#>    dry_end dry_mid low_n low_e low_start low_end low_mid low_magnitude
#> 1       NA      NA    NA    NA        NA      NA      NA            NA
#> 2       NA      NA    NA    NA        NA      NA      NA            NA
#> 3       NA      NA    NA    NA        NA      NA      NA            NA
#> 4       NA      NA    NA    NA        NA      NA      NA            NA
#> 5       NA      NA    NA    NA        NA      NA      NA            NA
#> 6       NA      NA    NA    NA        NA      NA      NA            NA
#> 7       NA      NA    NA    NA        NA      NA      NA            NA
#> 8       NA      NA    NA    NA        NA      NA      NA            NA
#> 9       NA      NA    NA    NA        NA      NA      NA            NA
#> 10      NA      NA    NA    NA        NA      NA      NA            NA
#> 11      NA      NA    NA    NA        NA      NA      NA            NA
#> 12      NA      NA    NA    NA        NA      NA      NA            NA
#> 13      NA      NA    NA    NA        NA      NA      NA            NA
#> 14      NA      NA    NA    NA        NA      NA      NA            NA
#>    low_severity high_n high_e high_start high_end high_mid e_above3xq50
#> 1            NA     NA     NA         NA       NA       NA           NA
#> 2            NA     NA     NA         NA       NA       NA           NA
#> 3            NA     NA     NA         NA       NA       NA           NA
#> 4            NA     NA     NA         NA       NA       NA           NA
#> 5            NA     NA     NA         NA       NA       NA           NA
#> 6            NA     NA     NA         NA       NA       NA           NA
#> 7            NA     NA     NA         NA       NA       NA           NA
#> 8            NA     NA     NA         NA       NA       NA           NA
#> 9            NA     NA     NA         NA       NA       NA           NA
#> 10           NA     NA     NA         NA       NA       NA           NA
#> 11           NA     NA     NA         NA       NA       NA           NA
#> 12           NA     NA     NA         NA       NA       NA           NA
#> 13           NA     NA     NA         NA       NA       NA           NA
#> 14           NA     NA     NA         NA       NA       NA           NA
#>    e_above5xq50 e_above7xq50 volume vol_z min min_z min_doy min_7day min_7day_z
#> 1            NA           NA     NA    NA  NA    NA      NA       NA         NA
#> 2            NA           NA     NA    NA  NA    NA      NA       NA         NA
#> 3            NA           NA     NA    NA  NA    NA      NA       NA         NA
#> 4            NA           NA     NA    NA  NA    NA      NA       NA         NA
#> 5            NA           NA     NA    NA  NA    NA      NA       NA         NA
#> 6            NA           NA     NA    NA  NA    NA      NA       NA         NA
#> 7            NA           NA     NA    NA  NA    NA      NA       NA         NA
#> 8            NA           NA     NA    NA  NA    NA      NA       NA         NA
#> 9            NA           NA     NA    NA  NA    NA      NA       NA         NA
#> 10           NA           NA     NA    NA  NA    NA      NA       NA         NA
#> 11           NA           NA     NA    NA  NA    NA      NA       NA         NA
#> 12           NA           NA     NA    NA  NA    NA      NA       NA         NA
#> 13           NA           NA     NA    NA  NA    NA      NA       NA         NA
#> 14           NA           NA     NA    NA  NA    NA      NA       NA         NA
#>    min_7day_doy min_30day min_30day_z min_30day_doy max max_z max_doy
#> 1            NA        NA          NA            NA  NA    NA      NA
#> 2            NA        NA          NA            NA  NA    NA      NA
#> 3            NA        NA          NA            NA  NA    NA      NA
#> 4            NA        NA          NA            NA  NA    NA      NA
#> 5            NA        NA          NA            NA  NA    NA      NA
#> 6            NA        NA          NA            NA  NA    NA      NA
#> 7            NA        NA          NA            NA  NA    NA      NA
#> 8            NA        NA          NA            NA  NA    NA      NA
#> 9            NA        NA          NA            NA  NA    NA      NA
#> 10           NA        NA          NA            NA  NA    NA      NA
#> 11           NA        NA          NA            NA  NA    NA      NA
#> 12           NA        NA          NA            NA  NA    NA      NA
#> 13           NA        NA          NA            NA  NA    NA      NA
#> 14           NA        NA          NA            NA  NA    NA      NA
#>  [ reached 'max' / getOption("max.print") -- omitted 1126 rows ]
#> 
#> [[2]]
#> # A tibble: 2,660 × 5
#>    flow_site_id start_date end_date   parameter  value
#>    <chr>        <date>     <date>     <chr>      <dbl>
#>  1 031031       2010-01-01 2019-12-31 max       15.9  
#>  2 2008         2010-01-01 2019-12-31 max       84.7  
#>  3 2024         2010-01-01 2019-12-31 max        5.00 
#>  4 2050         2010-01-01 2019-12-31 max       17.3  
#>  5 254340011    2010-01-01 2019-12-31 max       14.4  
#>  6 2859TH       2010-01-01 2019-12-31 max        0.653
#>  7 4023         2010-01-01 2019-12-31 max       10.7  
#>  8 4036         2010-01-01 2019-12-31 max        1.48 
#>  9 4175         2010-01-01 2019-12-31 max        3.86 
#> 10 432800       2010-01-01 2019-12-31 max        2.33 
#> # ℹ 2,650 more rows
#> 

## Calculate a suite of long-term and seasonal flow statistics for all sites listed in site_col, and standardise QXX flow statistics using flow data specified in ref_col:
#calc_flowstats(data = data,
#               site_col = "flow_site_id",
#               date_col = "date",
#               flow_col = "HistoricalFlow",
#               ref_col = "NaturalisedFlow")